Made DateSlider a component
